Cantos Espirituales
Now you will be able to have the Cantos Espirituales hymnbook in digital media, with which you will be able to carry this tool at any time and place.

You will find all the hymns in the hymnbook and you can choose how you want to see it, modifying the type of font, font size and even the color, to make reading the hymns a more enjoyable experience.

We hope that it will be for the benefit and edification of the body of Christ, one more instrument to praise and glorify the name of God.